chore(headless): extract core search-box controller
