+++ packages/headless/src/controllers/core/search-box/headless-core-search-box-options.ts
@@ -5,8 +5,8 @@ import {
   StringValue,
   RecordValue,
 } from '@coveo/bueno';
-import {SuggestionHighlightingOptions} from '../../utils/highlight';
-import {requiredNonEmptyString} from '../../utils/validate-payload';
+import {SuggestionHighlightingOptions} from '../../../utils/highlight';
+import {requiredNonEmptyString} from '../../../utils/validate-payload';
 
 export interface SearchBoxOptions {
   /**
